Allow setting options.cors to a function in order to set Origin dynamically
Sometimes it is necessary to set the Access-Control-Allow-Origin header
based on the included request headers, for instance the Origin header.

Currently, this is impossible since the cors option must be specified
as a string ahead of time. This change allows a function to be used
instead, which can compute a new value based on the request/response
pair.